requirements.txt
allure-pytest
pytest==8.3.2
faker==27.0.0
requests